requirements.txt
streamlit>=1.20
pillow
rapidfuzz
pandas
requests
